comment & related questions

+++ src/main/java/cd/zzh/community/controller/PublishController.java
@@ -22,7 +22,7 @@ public class PublishController {
     private QuestionServive questionServive;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model){
         QuestionDTO question = questionServive.getById(id);
 
@@ -43,7 +43,7 @@ public class PublishController {
             @RequestParam("title") String title,
             @RequestParam("description") String description,
             @RequestParam("tag") String tag,
-            @RequestParam("id") Integer id,
+            @RequestParam("id") Long id,
             HttpServletRequest request,
             Model model){
